Relocate docsLink to core/utils
